Relocate matcher test utils
