+++ central/platform/reprocessor/reprocessor_impl_test.go
@@ -11,7 +11,6 @@ import (
 	v1 "github.com/stackrox/rox/generated/api/v1"
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/concurrency"
-	"github.com/stackrox/rox/pkg/fixtures"
 	"github.com/stackrox/rox/pkg/sac"
 	"github.com/stretchr/testify/suite"
 	"go.uber.org/mock/gomock"
@@ -39,7 +38,7 @@ func (s *platformReprocessorImplTestSuite) SetupTest() {
 	s.alertDatastore = alertDSMocks.NewMockDataStore(s.mockCtrl)
 	s.configDatastore = configDatastoreMocks.NewMockDataStore(s.mockCtrl)
 	s.deploymentDatastore = deploymentDSMocks.NewMockDataStore(s.mockCtrl)
-	s.matcher = fixtures.GetPlatformMatcherWithDefaultPlatformComponentConfig(s.mockCtrl)
+	s.matcher = platformmatcher.GetTestPlatformMatcherWithDefaultPlatformComponentConfig(s.mockCtrl)
 
 	s.reprocessor = &platformReprocessorImpl{
 		alertDatastore:      s.alertDatastore,
